% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pl_data}
\alias{pl_data}
\title{Match data for 2000/2001 to 2017/2018}
\format{A data frame with 6840 rows and 103 variables:
\describe{
  \item{Div}{Division, this is unnecessary for this data}
  \item{Date}{Match Date (dd/mm/yy)}
  \item{Time}{Time of match kick off}
  \item{HomeTeam}{Home Team}
  \item{AwayTeam}{Away Team}
  \item{FTHG and HG}{Full Time Home Team Goals}
  \item{FTAG and AG}{Full Time Away Team Goals}
  \item{FTR and Res}{Full Time Result (H=Home Win, D=Draw, A=Away Win)}
  \item{HTHG}{Half Time Home Team Goals}
  \item{HTAG}{Half Time Away Team Goals}
  \item{HTR}{Half Time Result (H=Home Win, D=Draw, A=Away Win)}
  \item{Attendance}{Crowd Attendance}
  \item{Referee}{Match Referee}
  \item{HS}{Home Team Shots}
  \item{AS}{Away Team Shots}
  \item{HST}{Home Team Shots on Target}
  \item{AST}{Away Team Shots on Target}
  \item{HHW}{Home Team Hit Woodwork}
  \item{AHW}{Away Team Hit Woodwork}
  \item{HC}{Home Team Corners}
  \item{AC}{Away Team Corners}
  \item{HF}{Home Team Fouls Committed}
  \item{AF}{Away Team Fouls Committed}
  \item{HFKC}{Home Team Free Kicks Conceded}
  \item{AFKC}{Away Team Free Kicks Conceded}
  \item{HO}{Home Team Offsides}
  \item{AO}{Away Team Offsides}
  \item{HY}{Home Team Yellow Cards}
  \item{AY}{Away Team Yellow Cards}
  \item{HR}{Home Team Red Cards}
  \item{AR}{Away Team Red Cards}
  \item{HBP}{Home Team Bookings Points (10 = yellow, 25 = red)}
  \item{ABP}{Away Team Bookings Points (10 = yellow, 25 = red)}
  }}
\usage{
pl_data
}
\description{
A dataset containing information on all matches from
the seasons from 2000 until 2018, and their team colours
}
\keyword{datasets}
