% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_stats_data.R
\name{PL_stats_data}
\alias{PL_stats_data}
\title{Function to calculate stats for Premier League teams}
\usage{
PL_stats_data(
  team_data,
  date_from,
  date_to,
  x_stat,
  y_stat,
  teams,
  x_stat_by = NA,
  y_stat_by = NA
)
}
\arguments{
\item{team_data}{Data Frame in the form output by read_PL_data}

\item{date_from}{Start date}

\item{date_to}{End date}

\item{x_stat}{The statistic output for the x axis. See details.}

\item{y_stat}{The statistic output for the y axis. See details.}

\item{teams}{The teams to output.}

\item{x_stat_by}{The aggregation for the x stat. See details.}

\item{y_stat_by}{The aggregation for the y stat. See details.}
}
\description{
This function outputs stats for given dates
}
