% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pl_data}
\alias{pl_data}
\title{Match data for 2000/2001 to 2017/2018}
\format{A data frame with 6840 rows and 103 variables:
\describe{
  \item{Div}{Division, this is unnecessary for this data}
  \item{Date}{Date of match}
  \item{HomeTeam}{Home Team}
  \item{AwayTeam}{Away Team}
  }}
\usage{
pl_data
}
\description{
A dataset containing information on all matches from
the seasons from 2000 until 2018, and their team colours
}
\keyword{datasets}
